requirements.txt
spotipy>=2.23.0
scikit-learn>=1.3.0
numpy>=1.24.0
python-dotenv>=1.0.0
musicbrainzngs>=0.7.1
requests>=2.31.0
